requirements.txt
streamlit
langchain 
langchain-openai 
langchain-core 
langchain-community
langgraph
chromadb
langchain-chroma
pypdf
langchain-tavily
python-dotenv
IPython